refactor(dimension): improve dimension validation and exception handling in related classes
