+++ src/App/Sin/Scuola/Controllers/ClassiController.php
@@ -6,7 +6,6 @@ namespace App\Scuola\Controllers;
 
 use App\Scuola\Models\AddStudentAction;
 use App\Scuola\Models\Classe;
-use App\Scuola\Requests\AddStudentRequest;
 use Domain\Nomadelfia\Persona\Models\Persona;
 use Illuminate\Http\Request;
