+++ src/App/Sin/Scuola/Exceptions/BadDimensionException.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace App\Scuola\Exceptions;
 
-use App\Scuola\Exceptions\SchoolException;
-
 final class BadDimensionException extends SchoolException
 {
     public static function isNotValid(string $dimension, string $msg): self
