+++ src/App/Sin/Scuola/Exceptions/CouldNotAssignAlunno.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace App\Scuola\Exceptions;
 
-use App\Scuola\Exceptions\SchoolException;
-
 final class CouldNotAssignAlunno extends SchoolException
 {
     public static function isNotValidAnno(string $anno): self
