+++ src/App/Sin/Scuola/Models/Anno.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Scuola\Models;
 
-use App\Nomadelfia\Exceptions\SchoolException;
+use App\Scuola\Exceptions\SchoolException;
 use Carbon\Carbon;
 use Domain\Nomadelfia\Persona\Models\Persona;
 use Illuminate\Database\Eloquent\Builder;
